Refactoring: FSD архитектура
