--- src/main.scss
@@ -1,5 +1,0 @@
-// Везде* такие паддинги. Применил к компонену Container в MUI
-.MuiContainer-root {
-  padding-left: 24px;
-  padding-right: 24px;
-}
