+++ src/main.tsx
@@ -1,10 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
+import App from "./app/index.tsx";
 import { BrowserRouter } from "react-router-dom";
 import { store } from "./redux/store.ts";
 import { Provider } from "react-redux";
-import "./main.scss";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
