+++ src/pages/CasePage/components/AboutItem/AboutItem.scss
