+++ src/pages/MainPage/components/DesinersCarousel/DesinersCarousel.scss
