+++ src/pages/MainPage/components/DesinersCarousel/DesinersCarousel.tsx
