+++ src/pages/MainPage/components/DesinersCategories/DesinersCategories.scss
