+++ src/pages/MainPage/components/DesinersCategories/DesinersCategories.tsx
