+++ src/pages/MainPage/components/Feed/Feed.scss
