+++ src/pages/MainPage/components/Intro/Intro.scss
