+++ src/pages/MainPage/components/WorkCategories/WorkCategories.scss
