+++ src/pages/ProfilePage/components/index.ts
@@ -1,5 +1,4 @@
 import Work from "./Work/Work";
-import AboutItem from "../../../components/AboutItem/AboutItem";
 import EmptyData from "./EmptyData/EmptyData";
 import Info from "./Info/Info";
 import InfoAction from "./InfoAction/InfoAction";
@@ -7,6 +6,7 @@ import Portfolio from "./Portfolio/Portfolio";
 import Profile from "./Profile/Profile";
 import ProfileNav from "./ProfileNav/ProfileNav";
 import SocialIndicator from "./SocialIndicator/SocialIndicator";
+import AboutItem from "@/pages/CasePage/components/AboutItem/AboutItem";
 
 export {
   Work,
