+++ src/shared/Footer/Footer.scss
