+++ src/shared/Footer/__tests__/Footer.test.tsx
@@ -1,4 +1,4 @@
-import Footer from "../Footer/Footer";
+import Footer from "../Footer";
 import { render, screen } from "@testing-library/react";
 
 describe("Footer", () => {
