+++ src/shared/Preloader/Preloader.scss
