+++ src/shared/UI/MyAuthForm/MyAuthForm.scss
