+++ src/shared/UI/MyAuthForm/MyAuthForm.tsx
