+++ src/shared/UI/MyCheckBox/MyCheckBox.scss
