+++ src/shared/UI/MyCheckBox/MyCheckBox.tsx
