+++ src/shared/UI/MyInput/MyInput.scss
