+++ src/shared/UI/MyPopup/MyPopup.scss
