+++ src/shared/UI/MySwiper/MySwiper.scss
