+++ src/shared/UI/MySwiper/MySwiper.tsx
