+++ src/shared/UI/index.ts
@@ -1,0 +1,17 @@
+import MyAuthForm from "./MyAuthForm/MyAuthForm";
+import MyButton from "./MyButton/MyButton";
+import MyCheckBox from "./MyCheckBox/MyCheckBox";
+import MyDropDown from "./MyDropDown/MyDropDown";
+import MyInput from "./MyInput/MyInput";
+import MyPopup from "./MyPopup/MyPopup";
+import MySwiper from "./MySwiper/MySwiper";
+
+export {
+  MyAuthForm,
+  MyButton,
+  MyCheckBox,
+  MyDropDown,
+  MyInput,
+  MyPopup,
+  MySwiper,
+};
